move scripts to implement a better naming convention while importing
